requirements.txt
mysql-connector-python
Flask 
langchain_community